ENG-1784: Adding tablet server wide metrics for total, single and multi touch block cache usage.
Summary: Adding tablet server wide metrics for total, single touch and multi touch block caches. More detail available in https://yugabyte.atlassian.net/browse/ENG-1784

Test Plan: Add c test

Reviewers: karthik, kannan, sergei, amitanand

Reviewed By: amitanand

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2466

+++ src/yb/util/metrics.h
@@ -501,7 +501,7 @@ class MetricEntity : public RefCountedThreadSafe<MetricEntity> {
   // Set a particular attribute. Replaces any current value.
   void SetAttribute(const std::string& key, const std::string& val);
 
-  int num_metrics() const {
+  size_t num_metrics() const {
     std::lock_guard<simple_spinlock> l(lock_);
     return metric_map_.size();
   }
@@ -603,7 +603,7 @@ class MetricRegistry {
   void RetireOldMetrics();
 
   // Return the number of entities in this registry.
-  int num_entities() const {
+  size_t num_entities() const {
     std::lock_guard<simple_spinlock> l(lock_);
     return entities_.size();
   }
